==> 5_downsampled_data_merged/4a_consensus_qc.ipynb <==
# pycisTopic analysis

Cell downsampled dataset, using consensus peak regions.

import pycisTopic
pycisTopic.__version__

import warnings
warnings.filterwarnings('ignore')
warnings.simplefilter('ignore')

import pickle
import pandas as pd

import os
wdir = '/lustre1/project/stg_00002/lcb/fderop/data/20211024_hca_benchmark_libds_merged/'
os.chdir( wdir )

import glob
from collections import OrderedDict
filenames = glob.glob('fragments_postbap/*.sinto.mm.fragments.tsv.gz')
samples = [item.replace(".sinto.mm.fragments.tsv.gz", "") for item in filenames]
samples = [item.replace("fragments_postbap/", "") for item in samples]
fragments_dict = {samples[i]: filenames[i] for i in range(len(samples))}
fragments_dict = OrderedDict(sorted(fragments_dict.items()))
fragments_dict.keys()

fragments_dict

# load consensus peak dict


load a dictionary with paths to consensus files, generated in notebook 3

import glob
from collections import OrderedDict
consensus_peaks_dict = {}
key = 'libds_merged'
consensus_regions = 'consensus_peak_calling/'+key+'/'+key+'__consensus_regions.bed'


filenames = glob.glob('fragments_postbap/*.sinto.mm.fragments.tsv.gz')
samples = [item.replace(".sinto.mm.fragments.tsv.gz", "") for item in filenames]
samples = [item.replace("fragments_postbap/", "") for item in samples]
consensus_peaks_dict = {samples[i]: consensus_regions for i in range(len(samples))}
consensus_peaks_dict = OrderedDict(sorted(consensus_peaks_dict.items()))
consensus_peaks_dict.keys()

## Library QC metrics

from pycisTopic.qc import compute_qc_stats
import pyranges

! cp /staging/leuven/stg_00002/lcb/cflerin/analysis/pbmc_atac/analysis2/data_freeze_Jun2021/atac_qc_multiplet_merged/jupyter/biomart_annot.pickle .

import pickle
f_biomart = '/lustre1/project/stg_00002/lcb/fderop/data/20211013_hca_benchmark_libds/biomart_annot.pickle'
with open(f_biomart, 'rb') as f:
    annot = pickle.load(f)
annot

import ray
ray.shutdown()

metadata_bc_dict, profile_data_dict = compute_qc_stats(
        fragments_dict=fragments_dict,
        tss_annotation=annot,
        stats=['barcode_rank_plot', 'duplicate_rate', 'insert_size_distribution', 'profile_tss', 'frip'],
        label_list=None,
        path_to_regions=consensus_peaks_dict,
        n_cpu=24,
        valid_bc=None,
        n_frag=100,
        n_bc=None,
        tss_flank_window=2000,
        tss_window=50,
        tss_minimum_signal_window=100,
        tss_rolling_window=10,
        min_norm=0.1,
        remove_duplicates = True,
        )

f_qc_dir = 'pycistopic_consensus_peaks/qc__consensus_peaks'
if not os.path.exists(os.path.join(wdir, f_qc_dir)):
    os.makedirs(os.path.join(wdir, f_qc_dir))
    

with open(os.path.join(wdir, f_qc_dir, 'metadata.pickle'), 'wb') as f:
    pickle.dump(metadata_bc_dict, f)

with open(os.path.join(wdir, f_qc_dir, 'profile_data.pickle'), 'wb') as f:
    pickle.dump(profile_data_dict, f)

import seaborn as sns
