==> 5_downsampled_data_merged/4b_create_merged_cto_consensus.ipynb <==
# pycisTopic analysis

Cell downsampled dataset, using consensus peak regions.

import pycisTopic
pycisTopic.__version__

import warnings
warnings.filterwarnings('ignore')
warnings.simplefilter('ignore')

import pickle
import pandas as pd

import os
wdir = '/lustre1/project/stg_00002/lcb/fderop/data/20211024_hca_benchmark_libds_merged/'
os.chdir( wdir )

import glob
from collections import OrderedDict
filenames = glob.glob('fragments_postbap/*.sinto.mm.fragments.tsv.gz')
samples = [item.replace(".sinto.mm.fragments.tsv.gz", "") for item in filenames]
samples = [item.replace("fragments_postbap/", "") for item in samples]
fragments_dict = {samples[i]: filenames[i] for i in range(len(samples))}
fragments_dict = OrderedDict(sorted(fragments_dict.items()))
fragments_dict.keys()

fragments_dict

consensus_peaks_dict = {}
for key in fragments_dict.keys():
    consensus_peaks_dict[key] = '/lustre1/project/stg_00002/lcb/fderop/data/20211024_hca_benchmark_libds_merged/consensus_peak_calling/libds_merged/libds_merged__consensus_regions.bed'

## Create the cisTopic objects for each sample

f_qc_dir = 'pycistopic_consensus_peaks/qc__consensus_peaks'

with open(os.path.join(wdir, f_qc_dir, 'metadata.pickle'), 'rb') as f:
    metadata_bc_dict = pickle.load(f)
    
with open(os.path.join(wdir, f_qc_dir, 'profile_data.pickle'), 'rb') as f:
    profile_data_dict = pickle.load(f)

# use metadata tsvs to filter out our original cells

metadata_dict = {}
for key in fragments_dict.keys():
    f_meta = os.path.join('/lustre1/project/stg_00002/lcb/fderop/data/20211013_hca_benchmark_libds/pycistopic_screen',key+'_cell_data.tsv')
    metadata_dict[key] = pd.read_csv(f_meta, index_col=0, header=0, sep='\t')

bc_passing_filters = {}
for key in metadata_dict.keys():
    bc_passing_filters[key] = list(metadata_dict[key]['barcode'])

cells = 0
for key in bc_passing_filters.keys():
    print(f"{key}, {len(bc_passing_filters[key])}")
    cells = cells + len(bc_passing_filters[key])
print(f"total cells: {cells}")

cell counts are correct.

# create CTOs

from pycisTopic.cistopic_class import create_cistopic_object_from_fragments

#path_to_regions = '/staging/leuven/stg_00002/lcb/cbravo/SCREEN_ENCODE3/V2.hg38-rDHS-Unfiltered.blacklisted.bed'
path_to_blacklist = '/lustre1/project/stg_00002/lcb/fderop/data/20211013_hca_benchmark_libds/hg38_regions/hg38-blacklist.v2.bed'

f_cto_dir = 'pycistopic_consensus_peaks/cistopic_objs__mergedconsensus'
if not os.path.exists(os.path.join(wdir, f_cto_dir)):
    os.makedirs(os.path.join(wdir, f_cto_dir))

fragments_dict[key]

#Create all objects
import ray
ray.shutdown()

for key in fragments_dict.keys():
    f_out = os.path.join(wdir, f_cto_dir, key + '__cistopic_obj_mergedconsensus.pkl')
    if os.path.isfile(f_out):
        print(f"Skipping {key}")
        continue
        
    tmp_cto = create_cistopic_object_from_fragments(path_to_fragments=fragments_dict[key],
                                                    path_to_regions=consensus_peaks_dict[key],
                                                    path_to_blacklist=path_to_blacklist,
                                                    metrics=metadata_bc_dict[key], # why is this greyed out
                                                    valid_bc=bc_passing_filters[key],
                                                    n_cpu=6,
                                                    partition=20,
                                                    project=key)
    
    with open(f_out, 'wb') as f:
        pickle.dump(tmp_cto, f)
    
    print(f"DONE {key}")


#Create all objects
import ray
ray.shutdown()

# load objects into dict:


cistopic_obj_dict = {}
for key in bc_passing_filters.keys():
    f_out = os.path.join(wdir, f_cto_dir, key + '__cistopic_obj_mergedconsensus.pkl')
    with open(f_out, 'rb') as f:
        cistopic_obj_dict[key] = pickle.load(f)
    print(f"Loaded {key}")

import copy

f_out = os.path.join(wdir, f_cto_dir, 'libds_merged__cistopic_obj_mergedconsensus.pkl')
if os.path.isfile(f_out):
    print(f"Loading {f_out}")
    with open(f_out, 'rb') as f:
        cistopic_obj_dict['libds_merged'] = pickle.load(f)
else:
    # merge:
    ct_keys = list(cistopic_obj_dict.keys())
    cto_merged = copy.copy(cistopic_obj_dict[ct_keys[0]])
    cto_merged.merge([ cistopic_obj_dict[k] for k in ct_keys[1:] ], project='scATAC-seq_benchmark')
    with open(f_out, 'wb') as f:
        pickle.dump(cto_merged, f)
    
    cistopic_obj_dict['libds_merged'] = cto_merged

## Add cell annotations

# load metadata from the single samples
such as seurat cell type. this was written in notebook 5a

cistopic_obj_dict[key].cell_data

cistopic_obj_dict

keys = list(cistopic_obj_dict['libds_merged'].cell_data['sample_id'].unique())

metadata_dict = {}
for key in keys:
    f_meta = os.path.join('/lustre1/project/stg_00002/lcb/fderop/data/20211013_hca_benchmark_libds/pycistopic_screen',key+'_cell_data.tsv')
    metadata_dict[key] = pd.read_csv(f_meta, index_col=0, header=0, sep='\t')

metadata_dict.keys()

df_meta = pd.DataFrame()
for key in metadata_dict.keys():
    df_meta = pd.concat([df_meta, metadata_dict[key]], axis=0)

cistopic_obj_dict['libds_merged'].cell_data['seurat_cell_type'] = df_meta['seurat_cell_type']
cistopic_obj_dict['libds_merged'].cell_data['consensus_cell_type'] = df_meta['consensus_cell_type']

# save ctos

f_cto_dir = 'pycistopic_consensus_peaks/cistopic_objs__mergedconsensus'

if not os.path.exists(os.path.join(wdir, f_cto_dir)):
    os.makedirs(os.path.join(wdir, f_cto_dir))

for key in cistopic_obj_dict.keys():
    f_out = os.path.join(wdir, f_cto_dir, key + '__cistopic_obj_mergedconsensus_metadata_annotated.pkl')
    if os.path.isfile(f_out):
        print(f"Skipping {key}, already exists")
        continue
    with open(f_out, 'wb') as f:
        pickle.dump(cistopic_obj_dict[key], f)
    print(f"Generated and saved cistopic object for {key}")
