==> 5_downsampled_data_merged/6b_write_looms.ipynb <==
# pycisTopic analysis

Cell downsampled dataset, using consensus peak regions.

import pycisTopic
pycisTopic.__version__

import warnings
warnings.filterwarnings('ignore')
warnings.simplefilter('ignore')

import pickle
import pandas as pd

import os
wdir = '/lustre1/project/stg_00002/lcb/fderop/data/20211013_hca_benchmark_libds/'
os.chdir( wdir )

---

import glob
from collections import OrderedDict
filenames = glob.glob('fragments_postbap/*Hydrop*.sinto.mm.fragments.tsv.gz')
samples = [item.replace(".sinto.mm.fragments.tsv.gz", "") for item in filenames]
samples = [item.replace("fragments_postbap/", "") for item in samples]
fragments_dict = {samples[i]: filenames[i] for i in range(len(samples))}
fragments_dict = OrderedDict(sorted(fragments_dict.items()))
fragments_dict.keys()

# load ctos

f_cto_dir = 'pycistopic_consensus_peaks/cistopic_objs__consensus'

cistopic_obj_dict = {}
# load sample objects into dict:
for key in fragments_dict.keys():
    f_out = os.path.join(wdir, f_cto_dir, key + '__cistopic_obj_metadata_annotated.pkl')
    with open(f_out, 'rb') as f:
        cistopic_obj_dict[key] = pickle.load(f)
    print(f"Loaded {key}")

# extra: write looms to possibly re-do the label transfer (not done by cflerin)
i will test if this is worth it, for hydrop

f_dir = os.path.join(wdir, f_pseudo_dir)
f_consensus_dict = os.path.join(f_dir, 'consensus_peaks_dict.pkl')
if(os.path.isfile(f_consensus_dict)):
    print("consensus peaks dict already exists!")
else:
    print(f"file consensus peaks dict doesn't exist, writing")
    with open(f_consensus_dict, 'wb') as f:
        pickle.dump(consensus_peaks_dict, f)
    print(f"Finished {key}")

# write the consensus cell types into a dict.

meta1 = cistopic_obj_dict['VIB_Hydrop_1'].cell_data
meta2 = cistopic_obj_dict['VIB_Hydrop_2'].cell_data

meta_concat = pd.concat([meta1, meta2], axis=0, ignore_index=False)
meta_concat

len(cistopic_obj_dict['VIB_Hydrop_1'].cell_data) + len(cistopic_obj_dict['VIB_Hydrop_2'].cell_data)

so this is correct. now write it.

meta_concat.to_csv(
    os.path.join(wdir,'pycistopic_screen/filtered_cell_data.tsv'),
    sep='\t'
)

filename = os.path.join(wdir,'pycistopic_screen/filtered_cell_data.tsv')
print(filename)

f_crl_dir = 'pycistopic_consensus_peaks/cell_region_loom__screen'
if not os.path.exists(os.path.join(wdir, f_crl_dir)):
    os.makedirs(os.path.join(wdir, f_crl_dir))

write looms with fragment counts

import loompy as lp
for key in cistopic_obj_dict.keys():
    f_out = os.path.join(wdir, f_crl_dir, key + 'cell_region-all.loom')
    if os.path.isfile(f_out):
        print(f"Skipping {key} loom writing, already exists")
    else:
        print(f"started {key} loom writing")
        lp.create(
            filename = f_out,
            layers=cistopic_obj_dict[key].fragment_matrix,
            row_attrs={ 'Gene': cistopic_obj_dict[key].region_names }, 
            #col_attrs={ 'CellID': cistopic_obj_dict[key].cell_names }, 
            col_attrs={ 'CellID': [ x.split('-')[0] + '-' + x.split('-')[1]  for x in cistopic_obj_dict[key].cell_names ] },
            #file_attrs=attrs
        )
        print(f"Finished {key} loom writing")

write looms with accessibility

import loompy as lp
for key in cistopic_obj_dict.keys():
    f_out = os.path.join(wdir, f_crl_dir, key + 'cell_region-all.BINARYACCESSIBILITY.loom')
    if os.path.isfile(f_out):
        print(f"Skipping {key} loom writing, already exists")
    else:
        print(f"started {key} loom writing")
        lp.create(
            filename = f_out,
            layers=cistopic_obj_dict[key].binary_matrix,
            row_attrs={ 'Gene': cistopic_obj_dict[key].region_names }, 
            #col_attrs={ 'CellID': cistopic_obj_dict[key].cell_names }, 
            col_attrs={ 'CellID': [ x.split('-')[0] + '-' + x.split('-')[1]  for x in cistopic_obj_dict[key].cell_names ] },
            #file_attrs=attrs
        )
        print(f"Finished {key} loom writing")
