==> 5_downsampled_data_merged/8a_export_loom.ipynb <==
# pycisTopic analysis

Full dataset, using consensus peak regions.

import pycisTopic
pycisTopic.__version__

import warnings
warnings.filterwarnings('ignore')
warnings.simplefilter('ignore')

import pickle
import pandas as pd

import os
wdir = '/lustre1/project/stg_00002/lcb/fderop/data/20211215_hca_benchmark_github/5_downsampled_data_merged'
os.chdir( wdir )

# create output directory:
f_final_dir = os.path.join(wdir, 'downstream_analysis')
if not os.path.exists(f_final_dir):
    os.makedirs(f_final_dir)

import copy

## Load the cisTopic objects

f_cto_dir = 'pycistopic_consensus_peaks/cistopic_objs__mergedconsensus/'

cistopic_obj_dict = {}
for key in ['libds_merged']:
    f_cto = os.path.join(wdir, f_cto_dir, key + '__cistopic_obj_mergedconsensus_metadata_annotated_models_umap_tsne.pkl')
    if(os.path.isfile(f_cto)):
        with open(f_cto, 'rb') as f:
            cistopic_obj_dict[key] = pickle.load(f)
        print(f"Loaded filtered cistopic object {key}")
    else:
        print(f"file {f_cto} doesn't exist")

from collections import OrderedDict

cistopic_obj_dict['libds_merged'].cell_data['sample_id'].unique()

# display names of each sample. NOTE: the samples will be dislpayed in the order of this dict!
alias_dict = OrderedDict({
    "Broad_1": "BioRad ATAC 1",
    "Broad_2": "BioRad ATAC 2",
    "Stanford_1": "10x ATAC A1",
    "Stanford_2": "10x ATAC A2",
    "VIB_1": "10x ATAC B1",
    "VIB_2": "10x ATAC B2",
    "CNAG_1": "10x ATAC C1",
    "CNAG_2": "10x ATAC C2",
    "Broad_mito_1": "10x mtATAC 1",
    "Broad_mito_2": "10x mtATAC 2",
    "Sanger_1": "10x Multiome 1",
    "Sanger_2": "10x Multiome 2",
    "VIB_Hydrop_1": "Hydrop ATAC 1",
    "VIB_Hydrop_2": "Hydrop ATAC 2",
    "s3atac": "s3 ATAC",
    "merged": "Merged"
})

cistopic_obj_dict['libds_merged'].cell_data['alias'] = [alias_dict[x] for x in cistopic_obj_dict['libds_merged'].cell_data['sample_id']]

cistopic_obj_dict['libds_merged'].cell_data['sample_id'].unique()

for key in cistopic_obj_dict['libds_merged'].cell_data['sample_id'].unique():
    print(alias_dict[key])

#### Save/load

f_out = os.path.join(f_final_dir, 'region_bin_topics.pkl')
if os.path.isfile(f_out):
    print(f"Loading {f_out}")
    with open(f_out, 'rb') as f:
        region_bin_topics_dict = pickle.load(f)

f_out = os.path.join(f_final_dir, 'binarized_cell_topic.pkl')
if os.path.isfile(f_out):
    print(f"Loading {f_out}")
    with open(f_out, 'rb') as f:
        binarized_cell_topic_dict = pickle.load(f)

f_out = os.path.join(f_final_dir, 'imputed_acc_obj.pkl')
if os.path.isfile(f_out):
    print(f"Loading {f_out}")
    with open(f_out, 'rb') as f:
        imputed_acc_obj_dict = pickle.load(f)

f_out = os.path.join(f_final_dir, 'markers_dict.pkl')
if os.path.isfile(f_out):
    print(f"Loading {f_out}")
    with open(f_out, 'rb') as f:
        markers_dict_dict = pickle.load(f)

## Gene activity

import pyranges as pr
import requests
import pybiomart as pbm

### Infer gene activity

f_out = os.path.join(f_final_dir, 'gene_act_dict.pkl')
if os.path.isfile(f_out):
    print(f"Loading {f_out}")
    with open(f_out, 'rb') as f:
        gene_act_dict = pickle.load(f)
else:
    from pycisTopic.gene_activity import get_gene_activity
    from pycisTopic.diff_features import find_diff_features
    # For human
    dataset = pbm.Dataset(name='hsapiens_gene_ensembl',  host='http://www.ensembl.org')
    annot = dataset.query(attributes=['chromosome_name', 'start_position', 'end_position', 'strand', 'external_gene_name', 'transcription_start_site', 'transcript_biotype'])
    annot['Chromosome/scaffold name'] = 'chr' + annot['Chromosome/scaffold name'].astype(str)
    annot.columns=['Chromosome', 'Start', 'End', 'Strand', 'Gene','Transcription_Start_Site', 'Transcript_type']
    annot = annot[annot.Transcript_type == 'protein_coding']
    annot.Strand[annot.Strand == 1] = '+'
    annot.Strand[annot.Strand == -1] = '-'
    pr_annotation = pr.PyRanges(annot.dropna(axis = 0))
    pr_annotation

    # get chromosome sizes (hg38)
    target_url = 'http://hgdownload.cse.ucsc.edu/goldenPath/hg38/bigZips/hg38.chrom.sizes'
    chromsizes = pd.read_csv(target_url, sep='\t', header=None)
    chromsizes.columns = ['Chromosome', 'End']
    chromsizes['Start'] = [0]*chromsizes.shape[0]
    chromsizes = chromsizes.loc[:,['Chromosome', 'Start', 'End']]
    chromsizes = pr.PyRanges(chromsizes)
    chromsizes

    gene_act_dict = {}
    for key in cistopic_obj_dict.keys():
        print(key)
        gene_act, weights = get_gene_activity(
            imputed_acc_obj_dict[key], # Region-cell probabilities
            pr_annotation, # Gene annotation
            chromsizes, # Chromosome size
            use_gene_boundaries=True, # Whether to use the whole search space or stop when encountering another gene
            upstream=[1000, 100000], # Search space upstream. The minimum means that even if there is a gene right next to it 
                                     #these bp will be taken (1kbp here)
            downstream=[1000,100000], # Search space downstream
            distance_weight=True, # Whether to add a distance weight (an exponential function, the weight will decrease with distance)
            decay_rate=1, # Exponent for the distance exponential funciton (the higher the faster will be the decrease)
            extend_gene_body_upstream=10000, # Number of bp upstream immune to the distance weight (their value will be maximum for 
                                             #this weight)
            extend_gene_body_downstream=500, # Number of bp downstream immune to the distance weight
            gene_size_weight=False, # Whether to add a weights based on the length of the gene
            gene_size_scale_factor='median', # Dividend to calculate the gene size weigth. Default is the median value of all genes
                                             #in the genome
            remove_promoters=False, # Whether to remove promoters when computing gene activity scores
            average_scores=True, # Whether to divide by the total number of region assigned to a gene when calculating the gene 
                                 # activity score
            scale_factor=1, # Value to multiply for the final gene activity matrix
            extend_tss=[10,10], # Space to consider a promoter
            gini_weight = True, # Whether to add a gini index weigth. The more unique the region is, the higher this weight will be
            return_weights= True, # Whether to return the final weights
            project='Gene_activity') # Project name for the gene activity object
        gene_act_dict[key] = copy.copy(gene_act)

    with open(f_out, 'wb') as f:
        pickle.dump(gene_act_dict, f)
    print(f'saved {f_out}')

cistopic_obj_dict['libds_merged'].projections

## Export to loom

f_gene_loom_dir = os.path.join(f_final_dir, 'gene_act_loom')
if not os.path.exists(f_gene_loom_dir):
    os.makedirs(f_gene_loom_dir)

from pycisTopic.loom import (
    export_region_accessibility_to_loom,
    export_gene_activity_to_loom
)

### Gene activity

cistopic_obj_dict['libds_merged'].cell_data.columns

from ctxcore.genesig import Regulon

# generate a dummy regulon (required for export_gene_activity_to_loom):
phreg = Regulon(
        name='placeholder regulon',
        gene2weight={'phreg': 1.0},
        transcription_factor="phreg",
        gene2occurrence={"phreg": 1},
    )

for key in cistopic_obj_dict.keys():
    print(key)
    s = 'Merged' if key=='libds_merged' else alias_dict[key]
    f_out = os.path.join(f_gene_loom_dir, s + '__libds_gene_activity.loom')
    if os.path.exists(f_out):
        print(f"Skipping {f_out}: already exists.")
        continue
        
    export_gene_activity_to_loom(
        gene_activity_matrix = gene_act_dict[key],
        cistopic_obj = cistopic_obj_dict[key], 
        regulons = [phreg],
        # selected_cells = [ x.split('-')[0] + '-' + x.split('-')[1]  for x in cistopic_obj_dict[key].cell_names ], # this leaves a cell barcode of the format type 'TGCATGTCGCCGTTCCAAGA-21'
        # selected_cells = cistopic_obj_dict[key].projections['cell']['UMAP'].index.tolist(), # cflerin original 
        out_fname = f_out,
        cluster_annotation = ['consensus_cell_type', 'alias', 'fmx_sample'],
        cluster_markers = {'consensus_cell_type': markers_dict_dict[key], 'alias':{}, 'fmx_sample':{}},
        tree_structure = ('scATAC-seq_Benchmark', 'ATAC_library_downsampled', 'Gene_activity'),
        title = 'Gene activity from library-downsampled and sample-merged dataset',
        nomenclature = "hg38"
    )

### Region accessibility

f_region_loom_dir = os.path.join(f_final_dir, 'region_acc_loom')
if not os.path.exists(f_region_loom_dir):
    os.makedirs(f_region_loom_dir)

for key in cistopic_obj_dict.keys():
    print(key)
    s = 'Merged' if key=='libds_merged' else alias_dict[key]
    f_out = os.path.join(f_region_loom_dir, s + '__libds_region_accessibility.loom')

    if os.path.exists(f_out):
        print(f"Skipping {f_out}: already exists.")
        continue
    # Subset regions, we will use only regions in topics and DARs here to make it faster
    regions_in_topics = list(set(sum([region_bin_topics_dict[key][i].index.tolist() for i in region_bin_topics_dict[key].keys()],[])))
    regions_in_DARs = list(set(sum([markers_dict_dict[key][i].index.tolist() for i in markers_dict_dict[key].keys()],[])))
    # make sure we only take regions that actually exist in the accessibility matrix:
    selected_regions = list(set(regions_in_topics + regions_in_DARs).intersection(set(imputed_acc_obj_dict[key].feature_names)))

    # Export to loom
    export_region_accessibility_to_loom(
        accessibility_matrix = imputed_acc_obj_dict[key],
        cistopic_obj = cistopic_obj_dict[key], 
        binarized_topic_region = region_bin_topics_dict[key],
        binarized_cell_topic = binarized_cell_topic_dict[key],
        out_fname = f_out,
        selected_regions = selected_regions ,
        # selected_cells = [ x.split('-')[0] + '-' + x.split('-')[1]  for x in cistopic_obj_dict[key].cell_names ], # this leaves a cell barcode of the format type 'TGCATGTCGCCGTTCCAAGA-21'
        # selected_cells = cistopic_obj_dict[key].projections['cell']['UMAP'].index.tolist(), # cflerin original
        cluster_annotation = ['consensus_cell_type', 'alias', 'fmx_sample'],
        cluster_markers = {'consensus_cell_type': markers_dict_dict[key], 'alias':{}, 'fmx_sample':{}},
        tree_structure = ('scATAC', 'Region_accessibility'),
        title = f'Region accessibility from sample-merged dataset',
        nomenclature = "hg38"
    )

# Caution:
The export loom function contains a bug which adds the barcodes as a clustering, causing scope to crash on loading. We fix this below:

# open loom

import loompy as lp
import json
import copy
import glob
from collections import OrderedDict

exclude 'barcode' annotation

filenames = sorted(glob.glob(f_gene_loom_dir+'/*loom'))
samples = [item.replace("__libds_gene_activity.loom", "") for item in filenames]
samples = [item.replace("downstream_analysis/gene_act_loom/", "") for item in samples]
loom_dict = {samples[i]: filenames[i] for i in range(len(samples))}
loom_dict = OrderedDict(loom_dict)

loom_dict

for sample in loom_dict.keys():
    file = loom_dict[sample]
    print(f'opening {file}')
    loom = lp.connect(file, mode='r+', validate=False)

    metadata = json.loads(loom.attrs['MetaData'])
    new_metadata = copy.deepcopy(metadata)
    for item in new_metadata['annotations']:
        if item['name'] == 'barcode':
            barcode_index = new_metadata['annotations'].index(item)

    if 'barcode_index' in locals():
        print(f'\tbarcode index found: {barcode_index}')
        new_metadata['annotations'].pop(barcode_index)
        del barcode_index

        new_metadata_json = json.dumps(new_metadata)
        loom.attrs['MetaData'] = new_metadata_json
        loom.close()
        print(f'\tfixed {sample}')

    else:
        print('\tno barcode annotation detected')
        loom.close()

# remove the pesky sample_id from the non-merged looms

for sample in loom_dict.keys():
    file = loom_dict[sample]
    print(f'opening {file}')
    loom = lp.connect(file, mode='r+', validate=False)

    metadata = json.loads(loom.attrs['MetaData'])
    new_metadata = copy.deepcopy(metadata)
    for item in new_metadata['annotations']:
        if item['name'] == 'sample_id':
            barcode_index = new_metadata['annotations'].index(item)

    if 'barcode_index' in locals():
        print(f'\sample_id index found: {barcode_index}')
        new_metadata['annotations'].pop(barcode_index)
        del barcode_index

        new_metadata_json = json.dumps(new_metadata)
        loom.attrs['MetaData'] = new_metadata_json
        loom.close()
        print(f'\tfixed {sample}')

    else:
        print('\tno sample_id detected')
        loom.close()
    print('\n')

# open loom

import loompy as lp
import json
import copy
import glob
from collections import OrderedDict

exclude 'barcode' annotation

filenames = sorted(glob.glob(f_region_loom_dir+'/*loom'))
samples = [item.replace("__libds_region_accessibility.loom", "") for item in filenames]
samples = [item.replace("downstream_analysis/region_acc_loom/", "") for item in samples]
loom_dict = {samples[i]: filenames[i] for i in range(len(samples))}
loom_dict = OrderedDict(loom_dict)

loom_dict

for sample in loom_dict.keys():
    file = loom_dict[sample]
    print(f'opening {file}')
    loom = lp.connect(file, mode='r+', validate=False)

    metadata = json.loads(loom.attrs['MetaData'])
    new_metadata = copy.deepcopy(metadata)
    for item in new_metadata['annotations']:
        if item['name'] == 'barcode':
            barcode_index = new_metadata['annotations'].index(item)

    if 'barcode_index' in locals():
        print(f'\tbarcode index found: {barcode_index}')
        new_metadata['annotations'].pop(barcode_index)
        del barcode_index

        new_metadata_json = json.dumps(new_metadata)
        loom.attrs['MetaData'] = new_metadata_json
        loom.close()
        print(f'\tfixed {sample}')

    else:
        print('\tno barcode annotation detected')
        loom.close()

# remove the pesky sample_id from the non-merged looms

for sample in loom_dict.keys():
    file = loom_dict[sample]
    print(f'opening {file}')
    loom = lp.connect(file, mode='r+', validate=False)

    metadata = json.loads(loom.attrs['MetaData'])
    new_metadata = copy.deepcopy(metadata)
    for item in new_metadata['annotations']:
        if item['name'] == 'sample_id':
            barcode_index = new_metadata['annotations'].index(item)

    if 'barcode_index' in locals():
        print(f'\sample_id index found: {barcode_index}')
        new_metadata['annotations'].pop(barcode_index)
        del barcode_index

        new_metadata_json = json.dumps(new_metadata)
        loom.attrs['MetaData'] = new_metadata_json
        loom.close()
        print(f'\tfixed {sample}')

    else:
        print('\tno sample_id detected')
        loom.close()
    print('\n')
