==> 5_downsampled_data_merged/8b_upload_looms.ipynb <==
# upload

for file in /lustre1/project/stg_00002/lcb/fderop/data/20211215_hca_benchmark_github/5_downsampled_data_merged/downstream_analysis/gene_act_loom/*.loom
do
    echo "started uploading" $file
    curl -F 'UUID=scATAC-seq_Benchmark' -F 'file-type=Loom' -F 'file=@'$file https://scope.aertslab.org/upload/ | tee /dev/null
    echo "finished uploading" $file
done

for file in /lustre1/project/stg_00002/lcb/fderop/data/20211215_hca_benchmark_github/5_downsampled_data_merged/downstream_analysis/region_acc_loom/*.loom
do
    echo "started uploading" $file
    curl -F 'UUID=scATAC-seq_Benchmark' -F 'file-type=Loom' -F 'file=@'$file https://scope.aertslab.org/upload/ | tee /dev/null
    echo "finished uploading" $file
done
